+++ component/libfuse/libfuse2_handler.go
@@ -278,7 +278,7 @@ func libfuse2_init(conn *C.fuse_conn_info_t) (res unsafe.Pointer) {
 
 	C.populate_uid_gid()
 
-	log.Info("Libfuse::libfuse2_init : Kernel Caps : %d", conn.capable)
+	log.Info("Libfuse::libfuse2_init : Kernel FUSE version: %d.%d, Kernel Caps : 0x%x", conn.proto_major, conn.proto_minor, conn.capable)
 
 	if (conn.capable & C.FUSE_CAP_ASYNC_READ) != 0 {
 		log.Info("Libfuse::libfuse2_init : Enable Capability : FUSE_CAP_ASYNC_READ")
@@ -296,11 +296,14 @@ func libfuse2_init(conn *C.fuse_conn_info_t) (res unsafe.Pointer) {
 		conn.want |= C.FUSE_CAP_SPLICE_WRITE
 	}
 
-	// Max background thread on the fuse layer for high parallelism
-	conn.max_background = C.uint(fuseFS.maxFuseThreads)
+	// Max background requests on the fuse layer
+	conn.max_background = C.uint(fuseFS.maxBackground)
 
 	// While reading a file let kernel do readahed for better perf
 	conn.max_readahead = (4 * 1024 * 1024)
+
+	log.Info("Libfuse::libfuse2_init : want: 0x%x, max_readahead: %d, max_background: %d",
+		conn.want, conn.max_readahead, conn.max_background)
 
 	return nil
 }